+++ app/build.gradle
@@ -282,7 +282,7 @@ dependencies {
 
     // Markdown Parser
     implementation libs.markdown
-    testImplementation project(':core:testcommon')
+    testImplementation project(':core:testing')
 
     // Testing
     androidTestImplementation libs.androidx.junit
