+++ app/src/test/java/org/groundplatform/android/data/local/room/converter/ValueJsonConverterTest.kt
@@ -33,7 +33,7 @@ import org.groundplatform.domain.model.submission.TextTaskData
 import org.groundplatform.domain.model.task.MultipleChoice
 import org.groundplatform.domain.model.task.Option
 import org.groundplatform.domain.model.task.Task
-import org.groundplatform.testcommon.FakeDataGenerator
+import org.groundplatform.testing.FakeDataGenerator
 import org.json.JSONArray
 import org.junit.Test
 import org.junit.runner.RunWith
