+++ app/src/test/java/org/groundplatform/android/repository/SubmissionRepositoryTest.kt
@@ -32,7 +32,7 @@ import org.groundplatform.domain.model.submission.ValueDelta
 import org.groundplatform.domain.model.task.Task
 import org.groundplatform.domain.repository.LocationOfInterestRepositoryInterface
 import org.groundplatform.domain.repository.UserRepositoryInterface
-import org.groundplatform.testcommon.FakeDataGenerator
+import org.groundplatform.testing.FakeDataGenerator
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
