+++ app/src/test/java/org/groundplatform/android/ui/datacollection/tasks/AbstractTaskViewModelTest.kt
@@ -29,7 +29,7 @@ import org.groundplatform.android.FakeData.JOB
 import org.groundplatform.android.ui.datacollection.components.ButtonAction
 import org.groundplatform.domain.model.submission.TaskData
 import org.groundplatform.domain.model.submission.TextTaskData
-import org.groundplatform.testcommon.FakeDataGenerator.newTask
+import org.groundplatform.testing.FakeDataGenerator.newTask
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
