+++ core/testing/src/commonMain/kotlin/org/groundplatform/testing/FakeCall.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.testcommon
+package org.groundplatform.testing
 
 /**
  * Records calls to a fake method and allows overriding its behavior for testing.
