+++ core/testing/src/commonMain/kotlin/org/groundplatform/testing/FakeDataGenerator.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.testcommon
+package org.groundplatform.testing
 
 import org.groundplatform.domain.model.Survey
 import org.groundplatform.domain.model.User
