+++ core/testing/src/commonMain/kotlin/org/groundplatform/testing/FakeSurveyRepository.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.testcommon
+package org.groundplatform.testing
 
 import kotlin.collections.plus
 import kotlinx.coroutines.flow.Flow
@@ -38,7 +38,7 @@ class FakeSurveyRepository : SurveyRepositoryInterface {
   val onGetRemoteSurveyCall = FakeCall<String, Survey?> { id -> remoteSurveys.find { it.id == id } }
 
   override suspend fun saveSurvey(survey: Survey) {
-    offlineSurveys + survey
+    offlineSurveys = offlineSurveys + survey
   }
 
   override suspend fun getRemoteSurvey(surveyId: String): Survey? = onGetRemoteSurveyCall(surveyId)
