+++ settings.gradle
@@ -29,4 +29,4 @@ plugins {
 include ':app', ':e2eTest'
 include ':core:ui'
 include ':core:domain'
-include ':core:testcommon'
+include ':core:testing'
